refactor: split chat logic into hooks
