+++ src/components/chat/hooks/streaming-processor.ts
@@ -1,0 +1,431 @@
+import { cloudSync } from '@/services/cloud/cloud-sync'
+import { streamingTracker } from '@/services/cloud/streaming-tracker'
+import { logError } from '@/utils/error-handling'
+import type { Chat, Message } from '../types'
+
+export interface StreamingContext {
+  updatedChat: Chat
+  updatedMessages: Message[]
+  isFirstMessage: boolean
+  modelsLength: number
+  currentChatIdRef: React.MutableRefObject<string>
+  isStreamingRef: React.MutableRefObject<boolean>
+  thinkingStartTimeRef: React.MutableRefObject<number | null>
+  setIsThinking: (val: boolean) => void
+  setIsWaitingForResponse: (val: boolean) => void
+  updateChatWithHistoryCheck: (
+    setChats: React.Dispatch<React.SetStateAction<Chat[]>>,
+    chatSnapshot: Chat,
+    setCurrentChat: React.Dispatch<React.SetStateAction<Chat>>,
+    chatId: string,
+    newMessages: Message[],
+    immediate?: boolean,
+    isThinking?: boolean,
+  ) => void
+  setChats: React.Dispatch<React.SetStateAction<Chat[]>>
+  setCurrentChat: React.Dispatch<React.SetStateAction<Chat>>
+  setLoadingState: (s: 'idle' | 'loading') => void
+  storeHistory: boolean
+}
+
+export interface StreamingHandlers {
+  onAssistantMessageReady: (
+    assistantMessage: Message,
+    finalMessages: Message[],
+  ) => Promise<void>
+}
+
+export function getThinkingDuration(
+  thinkingStartTimeRef: React.MutableRefObject<number | null>,
+) {
+  const duration = thinkingStartTimeRef.current
+    ? (Date.now() - thinkingStartTimeRef.current) / 1000
+    : undefined
+  thinkingStartTimeRef.current = null
+  return duration
+}
+
+export async function processStreamingResponse(
+  response: Response,
+  ctx: StreamingContext,
+): Promise<Message | null> {
+  const reader = response.body?.getReader()
+  const decoder = new TextDecoder()
+  let assistantMessage: Message | null = {
+    role: 'assistant',
+    content: '',
+    timestamp: new Date(),
+  }
+
+  const streamingChatId = ctx.updatedChat.id
+  let rafId: number | ReturnType<typeof setTimeout> | null = null
+
+  try {
+    ctx.isStreamingRef.current = true
+    if (streamingChatId) streamingTracker.startStreaming(streamingChatId)
+
+    let thoughtsBuffer = ''
+    let isInThinkingMode = false
+    let isFirstChunk = true
+    let initialContentBuffer = ''
+    let sseBuffer = ''
+    let isUsingReasoningFormat = false
+
+    const scheduleStreamingUpdate = (isThinkingFlag: boolean) => {
+      if (rafId !== null) return
+      rafId =
+        typeof window !== 'undefined' &&
+        typeof window.requestAnimationFrame === 'function'
+          ? window.requestAnimationFrame(() => {
+              rafId = null
+              if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+                const chatId = ctx.currentChatIdRef.current
+                const messageToSave = assistantMessage as Message
+                const newMessages = [...ctx.updatedMessages, messageToSave]
+                ctx.updateChatWithHistoryCheck(
+                  ctx.setChats,
+                  ctx.updatedChat,
+                  ctx.setCurrentChat,
+                  chatId,
+                  newMessages,
+                  false,
+                  isThinkingFlag,
+                )
+              }
+            })
+          : setTimeout(() => {
+              if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+                const chatId = ctx.currentChatIdRef.current
+                const messageToSave = assistantMessage as Message
+                const newMessages = [...ctx.updatedMessages, messageToSave]
+                ctx.updateChatWithHistoryCheck(
+                  ctx.setChats,
+                  ctx.updatedChat,
+                  ctx.setCurrentChat,
+                  chatId,
+                  newMessages,
+                  false,
+                  isThinkingFlag,
+                )
+              }
+              rafId = null
+            }, 16)
+    }
+
+    while (true) {
+      const { done, value } = await reader!.read()
+      if (done || ctx.currentChatIdRef.current !== ctx.updatedChat.id) {
+        if (rafId !== null) {
+          if (
+            typeof window !== 'undefined' &&
+            typeof window.cancelAnimationFrame === 'function' &&
+            typeof rafId === 'number'
+          ) {
+            window.cancelAnimationFrame(rafId)
+          } else {
+            clearTimeout(rafId as ReturnType<typeof setTimeout>)
+          }
+          rafId = null
+        }
+
+        if (isFirstChunk && initialContentBuffer.trim()) {
+          assistantMessage = {
+            role: 'assistant',
+            content: initialContentBuffer.trim(),
+            timestamp: assistantMessage?.timestamp || new Date(),
+            isThinking: false,
+          }
+          if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+            const newMessages = [...ctx.updatedMessages, assistantMessage]
+            ctx.updateChatWithHistoryCheck(
+              ctx.setChats,
+              ctx.updatedChat,
+              ctx.setCurrentChat,
+              ctx.updatedChat.id,
+              newMessages,
+              false,
+              false,
+            )
+          }
+        } else if (isInThinkingMode && thoughtsBuffer.trim()) {
+          isInThinkingMode = false
+          ctx.setIsThinking(false)
+          const thinkingDuration = getThinkingDuration(ctx.thinkingStartTimeRef)
+
+          assistantMessage = isUsingReasoningFormat
+            ? {
+                role: 'assistant',
+                content: '',
+                thoughts: thoughtsBuffer.trim(),
+                timestamp: assistantMessage?.timestamp || new Date(),
+                isThinking: false,
+                thinkingDuration,
+              }
+            : {
+                role: 'assistant',
+                content: thoughtsBuffer.trim(),
+                timestamp: assistantMessage?.timestamp || new Date(),
+                isThinking: false,
+                thinkingDuration,
+              }
+          if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+            const chatId = ctx.currentChatIdRef.current
+            const newMessages = [...ctx.updatedMessages, assistantMessage]
+            ctx.updateChatWithHistoryCheck(
+              ctx.setChats,
+              ctx.updatedChat,
+              ctx.setCurrentChat,
+              chatId,
+              newMessages,
+              true,
+              false,
+            )
+          }
+        }
+        break
+      }
+
+      const chunk = decoder.decode(value)
+      sseBuffer += chunk
+      const lines = sseBuffer.split('\n')
+      sseBuffer = lines.pop() || ''
+
+      for (const line of lines.filter((l) => l.trim() !== '')) {
+        if (line === 'data: [DONE]' || !line.startsWith('data: ')) continue
+        try {
+          const jsonData = line.replace(/^data: /, '')
+          const json = JSON.parse(jsonData)
+
+          const hasReasoningContent =
+            'reasoning_content' in (json.choices?.[0]?.delta || {}) ||
+            'reasoning_content' in (json.choices?.[0]?.message || {})
+          const reasoningContent =
+            json.choices?.[0]?.message?.reasoning_content ||
+            json.choices?.[0]?.delta?.reasoning_content ||
+            ''
+          let content = json.choices?.[0]?.delta?.content || ''
+
+          if (
+            hasReasoningContent &&
+            !isUsingReasoningFormat &&
+            !isInThinkingMode
+          ) {
+            isUsingReasoningFormat = true
+            isInThinkingMode = true
+            ctx.setIsThinking(true)
+            ctx.thinkingStartTimeRef.current = Date.now()
+            assistantMessage = {
+              ...assistantMessage,
+              thoughts: reasoningContent || '',
+              isThinking: true,
+            }
+            if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+              const chatId = ctx.currentChatIdRef.current
+              const messageToSave = assistantMessage as Message
+              const newMessages = [...ctx.updatedMessages, messageToSave]
+              ctx.updateChatWithHistoryCheck(
+                ctx.setChats,
+                ctx.updatedChat,
+                ctx.setCurrentChat,
+                chatId,
+                newMessages,
+                false,
+                true,
+              )
+              requestAnimationFrame(() => {
+                ctx.setIsWaitingForResponse(false)
+              })
+            }
+            isFirstChunk = false
+            if (reasoningContent) thoughtsBuffer = reasoningContent
+            continue
+          } else if (isUsingReasoningFormat && hasReasoningContent) {
+            if (reasoningContent) thoughtsBuffer += reasoningContent
+            if (content && isInThinkingMode) {
+              isInThinkingMode = false
+              ctx.setIsThinking(false)
+              const thinkingDuration = getThinkingDuration(
+                ctx.thinkingStartTimeRef,
+              )
+              assistantMessage = {
+                ...assistantMessage,
+                thoughts: thoughtsBuffer.trim() || undefined,
+                content: content,
+                isThinking: false,
+                thinkingDuration,
+              }
+            } else if (reasoningContent) {
+              assistantMessage = {
+                ...assistantMessage,
+                thoughts: thoughtsBuffer,
+                isThinking: isInThinkingMode,
+              }
+            }
+            if (
+              (reasoningContent || content) &&
+              ctx.currentChatIdRef.current === ctx.updatedChat.id
+            ) {
+              scheduleStreamingUpdate(isInThinkingMode)
+            }
+            if (!content) continue
+          }
+
+          if (isUsingReasoningFormat && content) {
+            if (isInThinkingMode) {
+              isInThinkingMode = false
+              ctx.setIsThinking(false)
+              const thinkingDuration = getThinkingDuration(
+                ctx.thinkingStartTimeRef,
+              )
+              assistantMessage = { ...assistantMessage, thinkingDuration }
+            }
+            assistantMessage = {
+              ...assistantMessage,
+              content: (assistantMessage.content || '') + content,
+              isThinking: false,
+            }
+            if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+              scheduleStreamingUpdate(false)
+            }
+            continue
+          }
+
+          if (!isUsingReasoningFormat) {
+            if (isFirstChunk) {
+              initialContentBuffer += content
+              if (
+                initialContentBuffer.includes('<think>') ||
+                initialContentBuffer.length > 5
+              ) {
+                isFirstChunk = false
+                content = initialContentBuffer
+                initialContentBuffer = ''
+                if (content.includes('<think>')) {
+                  isInThinkingMode = true
+                  ctx.setIsThinking(true)
+                  ctx.thinkingStartTimeRef.current = Date.now()
+                  content = content.replace(/^[\s\S]*?<think>/, '')
+                  assistantMessage = {
+                    ...assistantMessage,
+                    isThinking: true,
+                    thoughts: '',
+                  }
+                  if (content) {
+                    thoughtsBuffer += content
+                    assistantMessage = {
+                      ...assistantMessage,
+                      thoughts: thoughtsBuffer,
+                    }
+                    content = ''
+                  }
+                  if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+                    const chatId = ctx.currentChatIdRef.current
+                    const messageToSave = assistantMessage as Message
+                    const newMessages = [...ctx.updatedMessages, messageToSave]
+                    ctx.updateChatWithHistoryCheck(
+                      ctx.setChats,
+                      ctx.updatedChat,
+                      ctx.setCurrentChat,
+                      chatId,
+                      newMessages,
+                      false,
+                      true,
+                    )
+                  }
+                }
+                requestAnimationFrame(() => {
+                  ctx.setIsWaitingForResponse(false)
+                })
+              } else {
+                continue
+              }
+            }
+          }
+
+          if (
+            content.includes('</think>') &&
+            isInThinkingMode &&
+            !isUsingReasoningFormat
+          ) {
+            isInThinkingMode = false
+            ctx.setIsThinking(false)
+            const thinkingDuration = getThinkingDuration(
+              ctx.thinkingStartTimeRef,
+            )
+            const parts = content.split('</think>')
+            const finalThoughts = (thoughtsBuffer + (parts[0] || '')).trim()
+            const remainingContent = parts.slice(1).join('')
+            assistantMessage = {
+              ...assistantMessage,
+              thoughts: finalThoughts || undefined,
+              isThinking: false,
+              thinkingDuration,
+            }
+            if (remainingContent.trim()) {
+              assistantMessage.content =
+                (assistantMessage.content || '') + remainingContent
+            }
+            if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+              scheduleStreamingUpdate(false)
+            }
+            continue
+          }
+
+          if (isInThinkingMode && !isUsingReasoningFormat) {
+            thoughtsBuffer += content
+            assistantMessage = {
+              ...assistantMessage,
+              thoughts: thoughtsBuffer,
+              isThinking: true,
+            }
+            if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+              scheduleStreamingUpdate(true)
+            }
+          } else if (!isInThinkingMode) {
+            if (!isUsingReasoningFormat) {
+              content = content.replace(/<think>|<\/think>/g, '')
+            }
+            if (content) {
+              assistantMessage = {
+                ...assistantMessage,
+                content: (assistantMessage.content || '') + content,
+                isThinking: false,
+              }
+              if (ctx.currentChatIdRef.current === ctx.updatedChat.id) {
+                scheduleStreamingUpdate(false)
+              }
+            }
+          }
+        } catch (error) {
+          logError('Failed to parse SSE line', error, {
+            component: 'streaming-processor',
+            metadata: { line },
+          })
+          continue
+        }
+      }
+    }
+  } finally {
+    ctx.setLoadingState('idle')
+    ctx.isStreamingRef.current = false
+    if (ctx.updatedChat.id) streamingTracker.endStreaming(ctx.updatedChat.id)
+    ctx.setIsThinking(false)
+    ctx.thinkingStartTimeRef.current = null
+    ctx.setIsWaitingForResponse(false)
+    if (
+      ctx.storeHistory &&
+      assistantMessage &&
+      (assistantMessage.content || assistantMessage.thoughts) &&
+      ctx.currentChatIdRef.current === ctx.updatedChat.id
+    ) {
+      cloudSync.backupChat(ctx.currentChatIdRef.current).catch((error) => {
+        logError('Failed to sync chat after streaming', error, {
+          component: 'streaming-processor',
+          action: 'finally',
+        })
+      })
+    }
+  }
+
+  return assistantMessage
+}
